Move chart to a different component
