+++ src/components/weather/TestChart.js
@@ -1,0 +1,38 @@
+import {
+  Chart as ChartJS,
+  CategoryScale,
+  LinearScale,
+  PointElement,
+  LineElement,
+  Title,
+  Tooltip,
+  Legend,
+} from 'chart.js';
+import { Line } from 'react-chartjs-2';
+ChartJS.register(
+  CategoryScale,
+  LinearScale,
+  PointElement,
+  LineElement,
+  Title,
+  Tooltip,
+  Legend
+);
+const TestChart = () => {
+  const options = {
+    responsive: true,
+  };
+  const data = {
+    labels: ['January', 'February', 'March', 'April', 'May', 'June', 'July'],
+    datasets: [{
+      label: 'My First Dataset',
+      data: [65, 59, 80, 81, 56, 55, 40],
+      fill: true,
+      borderColor: 'rgb(75, 192, 192)',
+      tension: 0.1
+    }],
+  };
+  return (<Line options={options} data={data} />);
+};
+
+export default TestChart;
