+++ src/components/weather/Weather.js
@@ -1,41 +1,10 @@
 import { useParams } from 'react-router-dom';
-import {
-  Chart as ChartJS,
-  CategoryScale,
-  LinearScale,
-  PointElement,
-  LineElement,
-  Title,
-  Tooltip,
-  Legend,
-} from 'chart.js';
-import { Line } from 'react-chartjs-2';
-ChartJS.register(
-  CategoryScale,
-  LinearScale,
-  PointElement,
-  LineElement,
-  Title,
-  Tooltip,
-  Legend
-);
+
+
 const Weather = () => {
-  const options = {
-    responsive: true,
-  };
-  const data = {
-    labels: ['January', 'February', 'March', 'April', 'May', 'June', 'July'],
-    datasets: [{
-      label: 'My First Dataset',
-      data: [65, 59, 80, 81, 56, 55, 40],
-      fill: true,
-      borderColor: 'rgb(75, 192, 192)',
-      tension: 0.1
-    }],
-  };
   const { city } = useParams(); 
   return (
-    <div className="weather">{city}<Line options={options} data={data} /></div>
+    <div className="weather">{city}</div>
     
   );
 };
